chore: factor `item_test` out of `test_common`
